Sytem-arduino/requirements_arduino.txt
# Assistente de Voz com Arduino - Dependências Python
SpeechRecognition==3.9.0
pyaudio==0.2.11
pyttsx3==2.90
pyserial==3.5
numpy==1.21.5
setuptools